Wrong datatype
Issue with the datatype for the ratio/product.

+++ MIST/pruned_tree.py
@@ -15,8 +15,8 @@ print("Python argv:", sys.argv)
 print("Current dir:", os.getcwd())
 print("Save folder:", save_folder)
 
-chain_ratio   = int(sys.argv[4])
-chain_product = int(sys.argv[5])
+chain_ratio   = float(sys.argv[4])
+chain_product = float(sys.argv[5])
 
 chain_trunc = 6
